# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  constants.AuthMethod:
    enum:
    - OIDC
    - SAML
    - SECRET_KEY
    type: string
    x-enum-varnames:
    - AuthMethodOIDC
    - AuthMethodSAML
    - AuthMethodSecretKey
...
  constants.IAMRoleType:
    enum:
    - platform
    - workspace
    - csp
    type: string
    x-enum-comments:
      RoleTypeCSP: CSP 역할
      RoleTypePlatform: 플랫폼 역할
      RoleTypeWorkspace: 워크스페이스 역할
    x-enum-varnames:
    - RoleTypePlatform
    - RoleTypeWorkspace
    - RoleTypeCSP
...
  model.AuthMethodType:
    enum:
    - OIDC
    - SAML
    - SECRET_KEY
    type: string
    x-enum-varnames:
    - AuthMethodOIDC
    - AuthMethodSAML
    - AuthMethodSecretKey
...
  model.CspAccount:
    properties:
      account_info:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      csp_type:
        description: aws, gcp, azure
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CspIdpConfig:
    properties:
      auth_method:
        allOf:
        - $ref: '#/definitions/model.AuthMethodType'
        description: OIDC, SAML, SECRET_KEY
      config:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      csp_account:
        $ref: '#/definitions/model.CspAccount'
      csp_account_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CspRole:
    properties:
      create_date:
        type: string
      created_at:
        type: string
      csp_account:
        $ref: '#/definitions/model.CspAccount'
      csp_account_id:
        description: CSP 계정 및 IDP 설정 참조 (신규 추가)
        type: integer
      csp_idp_config:
        $ref: '#/definitions/model.CspIdpConfig'
      csp_idp_config_id:
        type: integer
      csp_type:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      extended_config:
        additionalProperties: true
        type: object
      iam_identifier:
        type: string
      iam_role_id:
        type: string
      id:
        type: integer
      idp_identifier:
        type: string
      max_session_duration:
        type: integer
      name:
        type: string
      path:
        type: string
      permissions:
        items:
          type: string
        type: array
      permissions_boundary:
        type: string
      role_last_used:
        $ref: '#/definitions/model.RoleLastUsed'
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      updated_at:
        type: string
    type: object
...
  model.McmpApiCallRequest:
    properties:
      actionName:
        description: Target action name (operationId)
        type: string
      requestParams:
        allOf:
        - $ref: '#/definitions/model.McmpApiRequestParams'
        description: Parameters for the external API call
      serviceName:
        description: Target service name
        type: string
    required:
    - actionName
    - serviceName
    type: object
  model.McmpApiRequestParams:
    properties:
      body:
        description: Request body (accept any JSON structure) - Changed from json.RawMessage
          for swag compatibility
      pathParams:
        additionalProperties:
          type: string
        description: Parameters to replace in the resource path (e.g., {userId})
        type: object
      queryParams:
        additionalProperties:
          type: string
        description: Parameters to append as query string (?key=value)
        type: object
    type: object
...
  model.RoleLastUsed:
    properties:
      last_used_date:
        type: string
      region:
        type: string
    type: object
  model.RoleMaster:
    properties:
      children:
        items:
          $ref: '#/definitions/model.RoleMaster'
        type: array
      created_at:
        type: string
      csp_role_mappings:
        items:
          $ref: '#/definitions/model.RoleMasterCspRoleMapping'
        type: array
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/model.RoleMaster'
      parent_id:
        type: integer
      predefined:
        type: boolean
      role_subs:
        items:
          $ref: '#/definitions/model.RoleSub'
        type: array
      updated_at:
        type: string
    type: object
  model.RoleMasterCspRoleMapping:
    properties:
      auth_method:
        $ref: '#/definitions/constants.AuthMethod'
      createdAt:
        type: string
      cspRoles:
        description: 서비스 레이어에서 조합
        items:
          $ref: '#/definitions/model.CspRole'
        type: array
      description:
        type: string
      roleId:
        type: integer
    type: object
...
  model.RoleSub:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_id:
        type: integer
      role_type:
        $ref: '#/definitions/constants.IAMRoleType'
      updated_at:
        type: string
    type: object
...
  model.Tag:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      description:
        type: string
      email:
        description: Ignore Email for DB
        type: string
      enabled:
        description: Enabled status managed by Keycloak
        type: boolean
      firstName:
        description: Ignore FirstName for DB
        type: string
      id:
        description: DB에 저장되는 정보 (mcmp_users 테이블)
        type: integer
      kc_id:
        description: Keycloak User ID
        type: string
      lastName:
        description: Ignore LastName for DB
        type: string
      organization:
        description: Organization stored in Keycloak attributes
        type: string
      platform_roles:
        description: 관계 정의
        items:
          $ref: '#/definitions/model.RoleMaster'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/model.UserStatus'
        description: 사용자 계정 상태
      updated_at:
        type: string
      username:
        description: Keycloak 정보
        type: string
      workspace_roles:
        items:
          $ref: '#/definitions/model.RoleMaster'
        type: array
    type: object
...
  model.UserStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - WITHDRAWAL_REQUESTED
    - WITHDRAWN
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusWithdrawalRequested
    - UserStatusWithdrawn
...
host: localhost
info:
  contact: {}
  description: MC IAM Manager API Documentation
  title: MC IAM Manager API
  version: 0.6.2
paths:
...
  /api/mcmp-apis/call:
    post:
      consumes:
      - application/json
      description: Executes a defined MCMP API action with parameters structured in
        McmpApiCallRequest.
      operationId: mcmpApiCall
      parameters:
      - description: API Call Request
        in: body
        name: callRequest
        required: true
        schema:
          $ref: '#/definitions/model.McmpApiCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: External API Response (structure depends on the called API)
          schema:
            type: object
        "400":
          description: 'error: Invalid request body or parameters'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Service or action not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error or failed to call external API'
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: 'error: External API unavailable'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Call an external MCMP API action (Structured Request)
      tags:
      - McmpAPI
...
  /api/users/kc/{kcUserId}:
    get:
      consumes:
      - application/json
      description: Get user details by KcID
      operationId: getUserByKcID
      parameters:
      - description: User KcID
        in: path
        name: kcUserId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by KcID
      tags:
      - users
...
  /api/workspaces/temporary-credentials:
    post:
      consumes:
      - application/json
      description: Get temporary credentials for CSP
      operationId: mciamGetTemporaryCredentials
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get temporary credentials
      tags:
      - csp-credentials
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"